Drag fields from drag icon only
